refactor: TC-51 Move Address to the newly created utility folder
