+++ apps/api/src/main/java/dk/treecreate/api/shipmondo/Sender.java
@@ -1,5 +1,7 @@
 package dk.treecreate.api.shipmondo;
 
+import dk.treecreate.api.shipmondo.utility.Address;
+
 public class Sender {
    
     private Address address;
