+++ apps/api/src/main/java/dk/treecreate/api/shipmondo/utility/Address.java
@@ -1,4 +1,4 @@
-package dk.treecreate.api.shipmondo;
+package dk.treecreate.api.shipmondo.utility;
 
 public class Address {
